Move `NavUser` to `app-sidebar.tsx`
