--- web/src/components/nav/nav-user.tsx
@@ -1,148 +1,0 @@
-"use client";
-
-import { ChevronsUpDown } from "lucide-react";
-import Link from "next/link";
-
-import {
-  Avatar,
-  AvatarFallback,
-  AvatarImage,
-} from "@/src/components/ui/avatar";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuGroup,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuSub,
-  DropdownMenuSubContent,
-  DropdownMenuSubTrigger,
-  DropdownMenuTrigger,
-} from "@/src/components/ui/dropdown-menu";
-import {
-  SidebarMenu,
-  SidebarMenuButton,
-  SidebarMenuItem,
-} from "@/src/components/ui/sidebar";
-
-export type UserNavigationItem = {
-  name: string;
-  onClick?: () => void;
-  content?: React.ReactNode;
-  href?: string;
-  subItems?: UserNavigationItem[];
-};
-
-export type UserNavigationProps = {
-  user: {
-    name: string;
-    email: string;
-    avatar: string;
-  };
-  items: UserNavigationItem[];
-};
-
-export function NavUser({
-  user,
-  items,
-  isMobile,
-}: UserNavigationProps & { isMobile: boolean }) {
-  const initials = user.name
-    .split(" ")
-    .slice(0, 2)
-    .map((word) => word[0])
-    .join("")
-    .toUpperCase();
-
-  const renderMenuItem = (item: UserNavigationItem) => {
-    if (item.subItems?.length) {
-      return (
-        <DropdownMenuSub key={item.name}>
-          <DropdownMenuSubTrigger>
-            {item.content ?? item.name}
-          </DropdownMenuSubTrigger>
-          <DropdownMenuSubContent>
-            {item.subItems.map(renderMenuItem)}
-          </DropdownMenuSubContent>
-        </DropdownMenuSub>
-      );
-    }
-
-    if (item.href) {
-      return (
-        <DropdownMenuItem key={item.name} asChild>
-          <Link href={item.href}>{item.content ?? item.name}</Link>
-        </DropdownMenuItem>
-      );
-    }
-
-    return (
-      <DropdownMenuItem key={item.name} onClick={item.onClick}>
-        {item.content ?? item.name}
-      </DropdownMenuItem>
-    );
-  };
-
-  return (
-    <SidebarMenu>
-      <SidebarMenuItem>
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <SidebarMenuButton
-              size="lg"
-              className="data-[state=open]:bg-sidebar-accent data-[state=open]:text-sidebar-accent-foreground"
-            >
-              <Avatar className="h-8 w-8 rounded-lg">
-                <AvatarImage src={user.avatar} alt={user.name} />
-                <AvatarFallback className="rounded-lg">
-                  {initials}
-                </AvatarFallback>
-              </Avatar>
-              <div className="grid flex-1 text-left text-sm leading-tight">
-                <span className="truncate font-bold" title={user.name}>
-                  {user.name}
-                </span>
-                <span className="truncate text-xs" title={user.email}>
-                  {user.email}
-                </span>
-              </div>
-              <ChevronsUpDown className="ml-auto size-4" />
-            </SidebarMenuButton>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent
-            // No z-index: this dropdown portals into the `popover` overlay
-            // layer, which sits above the `agent` layer (the assistant window)
-            // by layer ORDER — so it opens above the window with no magic
-            // number. See components/ui/layer.tsx.
-            className="w-(--radix-dropdown-menu-trigger-width) min-w-56 rounded-lg"
-            side={isMobile ? "bottom" : "right"}
-            align="end"
-            sideOffset={4}
-          >
-            <DropdownMenuLabel className="p-0 font-normal">
-              <div className="flex items-center gap-2 px-1 py-1.5 text-left text-sm">
-                <Avatar className="h-8 w-8 rounded-lg">
-                  <AvatarImage src={user.avatar} alt={user.name} />
-                  <AvatarFallback className="rounded-lg">
-                    {initials}
-                  </AvatarFallback>
-                </Avatar>
-                <div className="grid flex-1 text-left text-sm leading-tight">
-                  <span className="truncate font-bold" title={user.name}>
-                    {user.name}
-                  </span>
-                  <span className="truncate text-xs" title={user.email}>
-                    {user.email}
-                  </span>
-                </div>
-              </div>
-            </DropdownMenuLabel>
-            <DropdownMenuSeparator />
-            <DropdownMenuGroup>{items.map(renderMenuItem)}</DropdownMenuGroup>
-          </DropdownMenuContent>
-        </DropdownMenu>
-      </SidebarMenuItem>
-    </SidebarMenu>
-  );
-}
